#597 Move common lock functionality to TokenLockBase.
